+++ bin/test.sh
@@ -30,11 +30,6 @@ $WP plugin activate user-switching
 $WP language core install it_IT
 $WP language plugin install user-switching it_IT
 
-# Run the functional tests:
-TEST_SITE_WP_URL=$WP_URL \
-	./vendor/bin/codecept run functional --steps "$1" \
-	|| ( TESTS_EXIT_CODE=$? && kill $PHP_SERVER_PROCESS_ID && exit $TESTS_EXIT_CODE )
-
 # Run the acceptance tests:
 TEST_SITE_WP_URL=$WP_URL \
 	./vendor/bin/codecept run acceptance --steps "$1" \
