--- tests/functional.suite.yml
@@ -1,24 +1,0 @@
-# Codeception Test Suite Configuration
-#
-# Suite for functional tests
-# Perform tests in browser using WPBrowser.
-
-actor: FunctionalTester
-modules:
-  enabled:
-    - WPDb
-    - WPBrowser
-  config:
-    WPDb:
-      dsn: 'mysql:host=%WP_TESTS_DB_HOST%;dbname=%WP_TESTS_DB_NAME%'
-      user: '%WP_TESTS_DB_USER%'
-      password: '%WP_TESTS_DB_PASS%'
-      url: '%TEST_SITE_WP_URL%'
-      dump: false
-      populate: false
-      cleanup: false
-    WPBrowser:
-      url: '%TEST_SITE_WP_URL%'
-      adminUsername: 'admin'
-      adminPassword: 'admin'
-      adminPath: '/wp-admin'
